reduce number of times hyper is instantiated
