+++ jdbc-core/src/test/java/com/salesforce/datacloud/jdbc/core/ConnectionQuerySettingsTest.java
@@ -15,15 +15,18 @@
  */
 package com.salesforce.datacloud.jdbc.core;
 
-import static org.assertj.core.api.Assertions.assertThat;
-
 import com.google.common.collect.Maps;
 import com.salesforce.datacloud.jdbc.hyper.HyperTestBase;
 import lombok.SneakyThrows;
 import lombok.val;
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
 
-public class ConnectionQuerySettingsTest extends HyperTestBase {
+import static com.salesforce.datacloud.jdbc.hyper.HyperTestBase.assertWithStatement;
+import static org.assertj.core.api.Assertions.assertThat;
+
+@ExtendWith(HyperTestBase.class)
+public class ConnectionQuerySettingsTest {
     @Test
     @SneakyThrows
     public void testLegacyQuerySetting() {
