+++ jdbc-core/src/test/java/com/salesforce/datacloud/jdbc/core/DataCloudStatementFunctionalTest.java
@@ -15,20 +15,25 @@
  */
 package com.salesforce.datacloud.jdbc.core;
 
-import static org.assertj.core.api.AssertionsForClassTypes.assertThatThrownBy;
-import static org.assertj.core.api.AssertionsForInterfaceTypes.assertThat;
-
 import com.salesforce.datacloud.jdbc.exception.DataCloudJDBCException;
 import com.salesforce.datacloud.jdbc.hyper.HyperServerConfig;
 import com.salesforce.datacloud.jdbc.hyper.HyperTestBase;
 import com.salesforce.datacloud.query.v3.DataCloudQueryStatus;
-import java.sql.ResultSet;
-import java.util.stream.Collectors;
 import lombok.SneakyThrows;
 import lombok.val;
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
 
-public class DataCloudStatementFunctionalTest extends HyperTestBase {
+import java.sql.ResultSet;
+import java.util.stream.Collectors;
+
+import static com.salesforce.datacloud.jdbc.hyper.HyperTestBase.assertWithConnection;
+import static com.salesforce.datacloud.jdbc.hyper.HyperTestBase.assertWithStatement;
+import static org.assertj.core.api.AssertionsForClassTypes.assertThatThrownBy;
+import static org.assertj.core.api.AssertionsForInterfaceTypes.assertThat;
+
+@ExtendWith(HyperTestBase.class)
+public class DataCloudStatementFunctionalTest {
     private static final HyperServerConfig configWithSleep =
             HyperServerConfig.builder().build();
