+++ jdbc-core/src/test/java/com/salesforce/datacloud/jdbc/core/JDBCLimitsTest.java
@@ -15,9 +15,6 @@
  */
 package com.salesforce.datacloud.jdbc.core;
 
-import static org.assertj.core.api.Assertions.assertThat;
-import static org.assertj.core.api.Assertions.assertThatExceptionOfType;
-
 import com.google.common.collect.Maps;
 import com.salesforce.datacloud.jdbc.exception.DataCloudJDBCException;
 import com.salesforce.datacloud.jdbc.hyper.HyperTestBase;
@@ -25,8 +22,15 @@ import lombok.SneakyThrows;
 import lombok.val;
 import org.apache.commons.lang3.StringUtils;
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
 
-public class JDBCLimitsTest extends HyperTestBase {
+import static com.salesforce.datacloud.jdbc.hyper.HyperTestBase.assertWithConnection;
+import static com.salesforce.datacloud.jdbc.hyper.HyperTestBase.assertWithStatement;
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.assertThatExceptionOfType;
+
+@ExtendWith(HyperTestBase.class)
+public class JDBCLimitsTest {
     @Test
     @SneakyThrows
     public void testLargeQuery() {
