+++ jdbc-core/src/test/java/com/salesforce/datacloud/jdbc/hyper/HyperTestBase.java
@@ -15,53 +15,34 @@
  */
 package com.salesforce.datacloud.jdbc.hyper;
 
-import static org.assertj.core.api.Assertions.assertThat;
-
 import com.google.common.collect.ImmutableMap;
 import com.salesforce.datacloud.jdbc.core.DataCloudConnection;
 import com.salesforce.datacloud.jdbc.core.DataCloudStatement;
 import com.salesforce.datacloud.jdbc.interceptor.AuthorizationHeaderInterceptor;
-import com.salesforce.datacloud.jdbc.interceptor.QueryIdHeaderInterceptor;
-import io.grpc.ManagedChannel;
 import io.grpc.ManagedChannelBuilder;
-import io.grpc.MethodDescriptor;
-import io.grpc.StatusRuntimeException;
-import io.grpc.inprocess.InProcessChannelBuilder;
-import java.sql.ResultSet;
-import java.util.ArrayList;
-import java.util.Iterator;
-import java.util.List;
-import java.util.Map;
-import java.util.Properties;
-import java.util.concurrent.atomic.AtomicInteger;
-import java.util.function.BiFunction;
 import lombok.SneakyThrows;
 import lombok.extern.slf4j.Slf4j;
 import lombok.val;
 import org.assertj.core.api.ThrowingConsumer;
-import org.grpcmock.GrpcMock;
-import org.grpcmock.junit5.InProcessGrpcMockExtension;
-import org.junit.jupiter.api.AfterAll;
-import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.Assertions;
-import org.junit.jupiter.api.BeforeAll;
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.TestInstance;
 import org.junit.jupiter.api.Timeout;
-import org.junit.jupiter.api.extension.ExtendWith;
-import salesforce.cdp.hyperdb.v1.CancelQueryParam;
-import salesforce.cdp.hyperdb.v1.HyperServiceGrpc;
-import salesforce.cdp.hyperdb.v1.QueryInfoParam;
-import salesforce.cdp.hyperdb.v1.QueryResultParam;
+import org.junit.jupiter.api.extension.BeforeAllCallback;
+import org.junit.jupiter.api.extension.ExtensionContext;
+
+import java.sql.ResultSet;
+import java.util.Map;
+import java.util.Properties;
+import java.util.concurrent.atomic.AtomicInteger;
+import java.util.concurrent.atomic.AtomicReference;
+
+import static org.assertj.core.api.Assertions.assertThat;
 
 @Slf4j
-@TestInstance(TestInstance.Lifecycle.PER_CLASS)
-@ExtendWith(InProcessGrpcMockExtension.class)
-public class HyperTestBase {
-    public HyperServerProcess instance;
+public class HyperTestBase implements BeforeAllCallback, ExtensionContext.Store.CloseableResource {
+    public static final AtomicReference<HyperServerProcess> instance = new AtomicReference<>();
 
     @SneakyThrows
-    public final void assertEachRowIsTheSame(ResultSet rs, AtomicInteger prev) {
+    public static void assertEachRowIsTheSame(ResultSet rs, AtomicInteger prev) {
         val expected = prev.incrementAndGet();
         val a = rs.getBigDecimal(1).intValue();
         assertThat(expected).isEqualTo(a);
@@ -69,7 +50,7 @@ public class HyperTestBase {
 
     @SafeVarargs
     @SneakyThrows
-    public final void assertWithConnection(
+    public static void assertWithConnection(
             ThrowingConsumer<DataCloudConnection> assertion, Map.Entry<String, String>... settings) {
         try (val connection =
                 getHyperQueryConnection(settings == null ? ImmutableMap.of() : ImmutableMap.ofEntries(settings))) {
@@ -79,7 +60,7 @@ public class HyperTestBase {
 
     @SafeVarargs
     @SneakyThrows
-    public final void assertWithStatement(
+    public static void assertWithStatement(
             ThrowingConsumer<DataCloudStatement> assertion, Map.Entry<String, String>... settings) {
         try (val connection = getHyperQueryConnection(
                         settings == null ? ImmutableMap.of() : ImmutableMap.ofEntries(settings));
@@ -88,153 +69,46 @@ public class HyperTestBase {
         }
     }
 
-    public DataCloudConnection getHyperQueryConnection() {
+    public static DataCloudConnection getHyperQueryConnection() {
         return getHyperQueryConnection(ImmutableMap.of());
     }
 
     @SneakyThrows
-    public DataCloudConnection getHyperQueryConnection(Map<String, String> connectionSettings) {
+    public static DataCloudConnection getHyperQueryConnection(Map<String, String> connectionSettings) {
         val properties = new Properties();
         properties.putAll(connectionSettings);
-        log.info("Creating connection to port {}", instance.getPort());
-        ManagedChannelBuilder<?> channel = ManagedChannelBuilder.forAddress("127.0.0.1", instance.getPort())
-                .usePlaintext();
-
+        val port = getInstancePort();
+        log.info("Creating connection to port {}", port);
+        ManagedChannelBuilder<?> channel = ManagedChannelBuilder.forAddress("127.0.0.1", port).usePlaintext();
         return DataCloudConnection.fromChannel(channel, properties);
     }
 
-    @SneakyThrows
-    @AfterAll
+    public static int getInstancePort() {
+        val hyper = instance.get();
+        Assertions.assertTrue((hyper != null) && hyper.isHealthy(), "Hyper wasn't started, failing test");
+        return hyper.getPort();
+    }
+
+    @Override
+    public void beforeAll(ExtensionContext context) {
+        if (instance.get() == null) {
+            instance.set(new HyperServerProcess());
+        }
+    }
+
+    @Override
     @Timeout(5_000)
-    public void afterAll() {
-        instance.close();
+    public void close() throws Throwable {
+        val hyper = instance.get();
+        if (hyper != null) {
+            hyper.close();
+        }
     }
 
-    @SneakyThrows
-    @BeforeAll
-    public void beforeAll() {
-        instance = new HyperServerProcess();
-    }
-
-    @BeforeEach
-    public void assumeHyperEnabled() {
-        Assertions.assertTrue((instance != null) && instance.isHealthy(), "Hyper wasn't started, failing test");
-    }
-
-    static class NoopTokenSupplier implements AuthorizationHeaderInterceptor.TokenSupplier {
+    public static class NoopTokenSupplier implements AuthorizationHeaderInterceptor.TokenSupplier {
         @Override
         public String getToken() {
             return "";
         }
-    }
-
-    private final List<HyperServerProcess> servers = new ArrayList<>();
-
-    private final List<ManagedChannel> channels = new ArrayList<>();
-
-    private final List<DataCloudConnection> connections = new ArrayList<>();
-
-    @AfterEach
-    public void cleanUpServers() {
-        servers.forEach(s -> {
-            try {
-                s.close();
-            } catch (Exception e) {
-                log.error("Failed to clean up hyper server process", e);
-            }
-        });
-
-        channels.forEach(c -> {
-            try{
-                c.shutdownNow();
-            } catch (Exception e) {
-                log.error("Failed to clean up channel", e);
-            }
-        });
-
-        connections.forEach(c -> {
-            try {
-                c.close();
-            } catch (Exception e) {
-                log.error("Failed to close data cloud connection", e);
-            }
-        });
-    }
-
-    protected DataCloudConnection getInterceptedClientConnection() {
-        return getInterceptedClientConnection(HyperServerConfig.builder().build());
-    }
-
-    @SneakyThrows
-    protected DataCloudConnection getInterceptedClientConnection(HyperServerConfig config) {
-
-        val server = config.start();
-        servers.add(server);
-
-        val auth = AuthorizationHeaderInterceptor.of(new HyperTestBase.NoopTokenSupplier());
-        val channel = ManagedChannelBuilder.forAddress("127.0.0.1", server.getPort())
-                .usePlaintext()
-                .intercept(auth)
-                .maxInboundMessageSize(64 * 1024 * 1024)
-                .build();
-
-        channels.add(channel);
-
-        val mocked = InProcessChannelBuilder.forName(GrpcMock.getGlobalInProcessName())
-                .intercept(auth)
-                .usePlaintext();
-
-        GrpcMock.resetMappings();
-
-        val stub = HyperServiceGrpc.newBlockingStub(channel);
-
-        proxyStreamingMethod(
-                stub,
-                HyperServiceGrpc.getExecuteQueryMethod(),
-                HyperServiceGrpc.HyperServiceBlockingStub::executeQuery);
-        proxyStreamingMethod(
-                stub,
-                HyperServiceGrpc.getGetQueryInfoMethod(),
-                HyperServiceGrpc.HyperServiceBlockingStub::getQueryInfo);
-        proxyStreamingMethod(
-                stub,
-                HyperServiceGrpc.getGetQueryResultMethod(),
-                HyperServiceGrpc.HyperServiceBlockingStub::getQueryResult);
-
-        val conn = DataCloudConnection.fromChannel(mocked, new Properties());
-        connections.add(conn);
-        return conn;
-    }
-
-    public static <ReqT, RespT> void proxyStreamingMethod(
-            HyperServiceGrpc.HyperServiceBlockingStub stub,
-            MethodDescriptor<ReqT, RespT> mock,
-            BiFunction<HyperServiceGrpc.HyperServiceBlockingStub, ReqT, Iterator<RespT>> method) {
-        GrpcMock.stubFor(GrpcMock.serverStreamingMethod(mock).willProxyTo((request, observer) -> {
-            final String queryId;
-            if (request instanceof QueryInfoParam) {
-                queryId = ((QueryInfoParam) request).getQueryId();
-            } else if (request instanceof QueryResultParam) {
-                queryId = ((QueryResultParam) request).getQueryId();
-            } else if (request instanceof CancelQueryParam) {
-                queryId = ((CancelQueryParam) request).getQueryId();
-            } else {
-                queryId = null;
-            }
-
-            val response = method.apply(
-                    queryId == null ? stub : stub.withInterceptors(new QueryIdHeaderInterceptor(queryId)), request);
-
-            try {
-                while (response.hasNext()) {
-                    observer.onNext(response.next());
-                }
-            } catch (StatusRuntimeException ex) {
-                observer.onError(ex);
-                return;
-            }
-
-            observer.onCompleted();
-        }));
     }
 }
